refactor: Move multi-session components, update imports
Rename/move multi-session UI components from settings/account and user into a new multi-session folder (manage-account(s), switch-account-submenu(-content| -item)). Update affected imports (account-settings, multi-session-plugin and component internals) to the new @/components/multi-session paths and fix UserView imports. Also add a note to AGENTS.md to avoid leaving nx build errors.
