+++ examples/start-shadcn-example/src/components/multi-session/manage-account.tsx
@@ -11,6 +11,7 @@ import {
 import type { Session, User } from "better-auth"
 import { ArrowLeftRight, LogOut, MoreHorizontal } from "lucide-react"
 import { toast } from "sonner"
+
 import { Button } from "@/components/ui/button"
 import { Card, CardContent } from "@/components/ui/card"
 import {
